Add correction factor so that all cells have roughly the same area

+++ mazes-for-programmers/kotlin/src/main/kotlin/io/github/ocirne/mazes/colorization/Strokes.kt
@@ -3,7 +3,7 @@ package io.github.ocirne.mazes.colorization
 import java.awt.BasicStroke
 import java.awt.Stroke
 
-class Strokes(val cellSize: Int) {
+class Strokes(val cellSize: Float) {
 
     fun getBasicWall(): Stroke {
         return BasicStroke(0.1f * cellSize)
